# web/docs/static/skills/relayhub-integration/references/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "RelayHub HTTP API",
    "version": "1.0.0",
    "description": "Shipped MVP routes. JSON bodies are limited to 1 MiB; unknown top-level request fields fail. IDs are opaque. HTTPS production origin. No tenant API is shipped; Go and TypeScript SDKs are available in the repository. Operations endpoints also run on the worker private listener. Standard errors use {error:{code,message}}; unknown routes use 404 not_found and unsupported methods 405 method_not_allowed. CORS is disabled by default; RELAYHUB_ALLOWED_ORIGINS enables exact-origin HTTP CORS and WebSocket origins. Allowed preflights return 204; unsupported methods/headers return 403 forbidden. Actual requests still authenticate. HTTP signature and CORS details: /docs/developer/auth.md."
  },
  "servers": [
    {
      "url": "https://relayhub.dungxbuif.com"
    }
  ],
...
  "paths": {
...
    "/api/v1/stream": {
      "get": {
        "operationId": "connectDurableStream",
        "summary": "Upgrade to the RelayHub v1 durable stream",
        "description": "Authenticate a short-lived stream:connect token before upgrade, require the relayhub.stream.v1 WebSocket subprotocol, then exchange frames defined by /docs/asyncapi.yaml. Application identity comes only from the token; NATS subjects and consumer state remain private.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "SocketToken": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Browser origin must exactly match RELAYHUB_ALLOWED_ORIGINS."
          },
          {
            "name": "Sec-WebSocket-Protocol",
            "in": "header",
            "required": true,
            "schema": {
              "const": "relayhub.stream.v1"
            },
            "description": "Required durable stream protocol version."
          },
          {
            "name": "Upgrade",
            "in": "header",
            "required": true,
            "schema": {
              "const": "websocket"
            }
          },
          {
            "name": "Connection",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sec-WebSocket-Version",
            "in": "header",
            "required": true,
            "schema": {
              "const": "13"
            }
          },
          {
            "name": "Sec-WebSocket-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Durable stream established with relayhub.stream.v1 selected."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    },
    "/ws": {
      "get": {
        "operationId": "connectWebSocket",
        "summary": "Upgrade to standard RFC 6455 WebSocket",
        "description": "URL query token requires ws:connect. Request relayhub.realtime.v2 for the app-isolated bidirectional channel, targeting and presence protocol; no subprotocol preserves legacy v1. Valid token with missing scope or disallowed Origin receives 403. Token expiry is checked at handshake only. Socket.IO is unsupported. See /docs/developer/websocket.md and the realtime frame schemas.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "SocketToken": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Browser origin must exactly match RELAYHUB_ALLOWED_ORIGINS; wildcard is not accepted."
          },
          {
            "name": "Sec-WebSocket-Protocol",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "relayhub.realtime.v2"
              ]
            },
            "description": "Selects the app-isolated realtime channel protocol. Omit only for legacy clients."
          },
          {
            "name": "Upgrade",
            "in": "header",
            "required": true,
            "description": "RFC 6455 upgrade header",
            "schema": {
              "type": "string"
            },
            "example": "websocket"
          },
          {
            "name": "Connection",
            "in": "header",
            "required": true,
            "description": "RFC 6455 connection upgrade",
            "schema": {
              "type": "string"
            },
            "example": "Upgrade"
          },
          {
            "name": "Sec-WebSocket-Version",
            "in": "header",
            "required": true,
            "description": "RFC 6455 version",
            "schema": {
              "type": "string"
            },
            "example": "13"
          },
          {
            "name": "Sec-WebSocket-Key",
            "in": "header",
            "required": true,
            "description": "Base64-encoded 16-byte random challenge",
            "schema": {
              "type": "string"
            },
            "example": "dGhlIHNhbXBsZSBub25jZQ=="
          }
        ],
        "responses": {
          "101": {
            "description": "Upgrade to standard RFC 6455 WebSocket",
            "headers": {
              "Upgrade": {
                "schema": {
                  "const": "websocket"
                },
                "description": "RFC 6455 upgrade."
              },
              "Connection": {
                "schema": {
                  "type": "string"
                },
                "description": "Upgrade."
              },
              "Sec-WebSocket-Accept": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6455 challenge response."
              }
            }
          },
          "400": {
            "description": "Standard WebSocket upgrade required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_upgrade",
                    "message": "A standard WebSocket upgrade is required."
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "description": "Missing ws:connect scope or rejected browser Origin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "scope": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The token requires ws:connect."
                      }
                    }
                  },
                  "origin": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "Origin is not allowed."
                      }
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Standard WebSocket upgrade required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_upgrade",
                    "message": "A standard WebSocket upgrade is required."
                  }
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-relayhub_admin",
        "description": "Secure HttpOnly browser Admin session. Non-safe methods also require X-RelayHub-CSRF."
      },
      "AppApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RelayHub-Api-Key",
        "description": "Application API key; must accompany timestamp and signature."
      },
      "AppSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RelayHub-Signature",
        "description": "Lowercase hex HMAC-SHA256 of timestamp + LF + method + LF + exact request target (path and query) + LF + lowercase SHA256 of exact body bytes. No trailing LF. Use the UTF-8 bytes of the entire issued HMAC secret; do not decode its prefix or suffix. See /docs/developer/auth.md."
      },
      "SocketToken": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Short-lived server-issued token with ws:connect or stream:connect scope. URL-encode it; omit query strings from logs."
      }
    },
...
    "responses": {
      "Error400": {
        "description": "The request is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "invalid_request",
                "message": "The request is invalid."
              }
            }
          }
        }
      },
      "Error401": {
        "description": "Authentication failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "Authentication failed."
              }
            }
          }
        }
      },
      "Error403": {
        "description": "The authenticated application cannot access this resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "forbidden",
                "message": "The authenticated application cannot access this resource."
              }
            }
          }
        }
      },
...
      "Error413": {
        "description": "Request body exceeds the 1 MiB limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "request_too_large",
                "message": "Request body exceeds the 1 MiB limit."
              }
            }
          }
        }
      },
      "Error500": {
        "description": "An internal error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "internal_error",
                "message": "An internal error occurred."
              }
            }
          }
        }
      },
      "Error503": {
        "description": "No function handler claimed this invocation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "function_unavailable",
                "message": "No function handler claimed this invocation."
              }
            }
          }
        }
      },
...
    }
  }
}